Add UserInfo Entity & API

+++ backend/qnaboard/src/main/java/com/ktds/azure/tenant/qnaboard/service/QnABoardService.java
@@ -4,9 +4,11 @@ import com.ktds.azure.tenant.qnaboard.dto.QnABoardDto;
 import com.ktds.azure.tenant.qnaboard.dto.QnABoardListDto;
 import com.ktds.azure.tenant.qnaboard.model.QnABoard;
 import com.ktds.azure.tenant.qnaboard.model.QnABoardState;
+import com.ktds.azure.tenant.qnaboard.model.UserInfo;
 import com.ktds.azure.tenant.qnaboard.repository.QnABoardRepository;
 import com.ktds.azure.tenant.qnaboard.util.QnABoardMapper;
 import jakarta.servlet.http.HttpSession;
+import org.apache.catalina.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -39,14 +41,26 @@ public class QnABoardService {
     }
 
     public QnABoardDto getQnABoardById(Long id, HttpSession httpSession){
+        UserInfo userInfo = getUserInfo(httpSession);
         QnABoard qnABoard = qnABoardRepository.findById(id).orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
 
-        return QnABoardMapper.toDto(qnABoard);
+        if (userInfo.getName().equals(qnABoard.getWriter())) {
+            return QnABoardMapper.toDto(qnABoard);
+        } else {
+            throw new RuntimeException("읽기 권한이 없습니다");
+        }
+
     }
 
     public QnABoardDto updateQnABoard(QnABoardDto qnABoardDto, HttpSession httpSession) {
+        UserInfo userInfo = getUserInfo(httpSession);
         QnABoard qnABoard = qnABoardRepository.findById(qnABoardDto.getId()).orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
-        return QnABoardMapper.toDto(qnABoardRepository.save(QnABoardMapper.toEntity(qnABoardDto)));
+
+        if (userInfo.getName().equals(qnABoard.getWriter())) {
+            return QnABoardMapper.toDto(qnABoardRepository.save(QnABoardMapper.toEntity(qnABoardDto)));
+        } else {
+            throw new RuntimeException("읽기 권한이 없습니다.");
+        }
     }
 
     public void updateQnABoardAnswer(QnABoardDto qnABoardDto) {
@@ -54,9 +68,21 @@ public class QnABoardService {
     }
 
     public void deleteQnABoardsByIds(QnABoardDto qnABoardDto, HttpSession httpSession) {
-        //QnABoard qnABoard = qnABoardRepository.findById(qnABoardDto.getId()).orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
-        // 음 .. 유저 권한을 어떻게 해야하지
-        qnABoardRepository.deleteById(qnABoardDto.getId());
+        UserInfo userInfo = getUserInfo(httpSession);
+        QnABoard qnABoard = qnABoardRepository.findById(qnABoardDto.getId()).orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
+
+        System.out.println(userInfo.getName());
+        System.out.println(qnABoard.getWriter());
+
+        if (userInfo.getName().equals(qnABoard.getWriter())) {
+            qnABoardRepository.deleteById(qnABoardDto.getId());
+        } else {
+            throw new RuntimeException("삭제 권한이 없습니다.");
+        }
+    }
+
+    public UserInfo getUserInfo(HttpSession httpSession){
+        return (UserInfo) httpSession.getAttribute("userInfo");
     }
 
 }
